requirements.txt
django
django-taggit
mysqlclient
requests
pillow
django-cors-headers